requirements.txt
PyMuPDF
requests
beautifulsoup4
youtube-transcript-api
faiss-cpu
sentence-transformers
